feat: add isotopes information
BREAKING CHANGE
